Travis: fix builds erroring out
The `composer.json` file requests several dependencies as `@dev` or `dev-develop`, but did not allow installing of unstable builds, which was causing the Travis builds to error out on the `composer install` without ever running any tests.

By allowing the installation of unstable versions of dependencies, this should be fixed.